+++ CHANGELOG.md
@@ -5,6 +5,15 @@ This package keep all common file for VTV Apps (Customer, Shop, Admin).
 
 ## [Unreleased]
 
+- Added: largestDiscount
+- hide payment info in order detail
+- remove status in transport handle history (user should not see this)
+- name for COMPLETED_ORDER_COD_SYSTEM
+
+## [1.0.0] - 14/07/2024
+
+- up size lazy list to 40
+
 ## [v0.0.1] - 29/03/2024
 
 ### Added